fix(console): fix application refresh token settings
