+++ packages/console/src/pages/ApplicationDetails/index.module.scss
@@ -37,4 +37,3 @@
     display: flex;
   }
 }
-
