+++ packages/console/src/pages/ApplicationDetails/utils.ts
@@ -1,9 +1,19 @@
-import { type ApplicationResponse } from '@logto/schemas';
+import {
+  customClientMetadataDefault,
+  type ApplicationResponse,
+  type Application,
+} from '@logto/schemas';
 
 export type ApplicationForm = Pick<
   ApplicationResponse,
   'name' | 'description' | 'oidcClientMetadata' | 'customClientMetadata' | 'isAdmin'
 >;
+
+const mapToUriFormatArrays = (value?: string[]) =>
+  value?.filter(Boolean).map((uri) => decodeURIComponent(uri)) ?? [];
+
+const mapToUriOriginFormatArrays = (value?: string[]) =>
+  value?.filter(Boolean).map((uri) => decodeURIComponent(uri.replace(/\/*$/, ''))) ?? [];
 
 export const applicationFormDataParser = {
   fromResponse: (data: ApplicationResponse): ApplicationForm => {
@@ -13,8 +23,29 @@ export const applicationFormDataParser = {
       name,
       description,
       oidcClientMetadata,
-      customClientMetadata,
+      customClientMetadata: {
+        ...customClientMetadataDefault,
+        ...customClientMetadata,
+      },
       isAdmin,
+    };
+  },
+  toUpdateApplicationData: (formData: ApplicationForm): Partial<Application> => {
+    return {
+      ...formData,
+      oidcClientMetadata: {
+        ...formData.oidcClientMetadata,
+        redirectUris: mapToUriFormatArrays(formData.oidcClientMetadata.redirectUris),
+        postLogoutRedirectUris: mapToUriFormatArrays(
+          formData.oidcClientMetadata.postLogoutRedirectUris
+        ),
+      },
+      customClientMetadata: {
+        ...formData.customClientMetadata,
+        corsAllowedOrigins: mapToUriOriginFormatArrays(
+          formData.customClientMetadata.corsAllowedOrigins
+        ),
+      },
     };
   },
 };
